refactor common into roc_env
